Restructure App And Add MVC Framework
