+++ app.js
@@ -9,19 +9,22 @@ const methodOverride = require("method-override");
 const mongoose = require("mongoose");
 const ejsMate = require("ejs-mate");
 const ExpressError = require("./utils/ExpressError");
-const {userSchema, feedbackSchema, Subjects} = require("./schema.js");
 const wrapAsync = require("./utils/wrapAsync.js");
 const Subject = require("./models/subject.js");
 const Feedback = require("./models/feedback.js");
-const Task = require("./models/task.js");
 const session = require("express-session");
 const passport = require("passport");
 const  localStrategy = require("passport-local");
 const User = require("./models/users");
 const flash = require("connect-flash");
-const Question = require("./models/newQues.js");
-const questionPaperRoutes = require("./routes/questionPapers");
 
+//ROUTES
+const questionPaperRoutes = require("./routes/questionPapers.js");
+const suggestedQuestion = require('./routes/suggestedQuestion.js');
+const taskRoutes = require("./routes/taskRoutes.js");
+const feedbackRoutes = require("./routes/feedbackRoutes.js");
+const subjectRoutes = require("./routes/subjectRoutes.js");
+const userRoutes = require("./routes/userRoutes.js");
 
 const mongoLink = "mongodb://127.0.0.1:27017/QBProject";
 
@@ -69,28 +72,6 @@ passport.deserializeUser(User.deserializeUser());
 
 
 
-//Validate user
-const validateUser = (req,res,next)=>{
-  let {error} = userSchema.validate(req.body);
-  if(error){
-    throw new ExpressError(404,error.details[0].message);
-  }else{
-    next();
-  }
-};
-
-//feedback validate middleware fn
-const validateFeedback = (req,res,next) => {
-  console.log(req.body);
-  const {error} = feedbackSchema.validate(req.body);
-  if(error) {
-    throw new ExpressError(404, error.details[0].message);
-  }
-  next();
-};
-
-
-
 //flash message
 app.use((req, res, next) => {
   res.locals.currUser = req.user;
@@ -101,52 +82,6 @@ app.use((req, res, next) => {
   next();
 });
 
-app.use("/", questionPaperRoutes);
-
-
-//isQuestionOwner
-const isQuestionOwner = async (req,res,next)=>{
-  let {id} = req.params;
-  let question = await Question.findById(id);
-  if (!question.username.equals(res.locals.currUser._id)) {
-    req.flash("error", "You are not the owner of the listings");
-    return res.redirect("/semester/newQuePage");
-  }
-  next();
-}
-
-
-//isTeacher/admin
-const isTeacher = async (req,res,next)=>{
-  if(req.user && req.user.role ===  "teacher"){
-    return next();
-  }
-  req.flash("error",
-    "Access denied: This feature is restricted to teachers only.");
-  return res.redirect("/");
-}
-
-
-
-//is Logged in middlewares
-const isLoggedIn = (req,res,next)=>{
-  if(!req.isAuthenticated()){
-      req.session.redirectUrl = req.originalUrl;
-      console.log(req.session.redirect);
-      req.flash("error","User Must Be Logged In!");
-      return res.redirect("/semester/login");
-  }
-  next();
-  
-}
-
-// route saving redirect middleware
-const saveReturnTo = (req,res,next)=>{
-  if (req.session.redirectUrl) {
-    res.locals.redirectUrl = req.session.redirectUrl;
-  }
-  next();
-}
 
 
 
@@ -162,128 +97,7 @@ app.get("/", async (req, res) => {
 });
 
 
-
-//display suggested question page
-app.get("/semester/newQuePage",wrapAsync(async (req,res)=>{
-  const question = await  Question.find({}).populate("username");
-  res.render("suggestedQuestion/suggestedQuesPage",{question});
-}))
-
-//Suggession Question form
-app.get("/semester/newQues",isLoggedIn,(req,res)=>{
-    res.render("suggestedQuestion/suggQuesForm");
-})
-
-//suggestion question post
-app.post("/semester/newQues",wrapAsync(async (req,res)=>{
-  let {subject,question} = req.body;
-
-  let newQuestion = new Question({
-    username:req.user._id,
-    subject:subject,
-    question:question
-  });
-  let ques = await newQuestion.save();
-  console.log(ques);
-  console.log(ques._id);
-  req.flash("success","You have contributed");
-  res.redirect("/semester/newQuePage");
-}))
-
-//suggested question edit get request
-
-app.get("/semester/newQues/:id/edit",isQuestionOwner,isLoggedIn,wrapAsync(async (req,res)=>{
-  let {id} = req.params;
-  const question = await  Question.findById(id).populate("username");
-  res.render("suggestedQuestion/suggQuesEdit",{question});
-}))
-
-
-//suggested question edit PUT request
-
-app.put("/semester/newQues/:id",isQuestionOwner,wrapAsync(async (req,res)=>{
-  let {id} = req.params;
-  let {subject,question}= req.body;
-  console.log(id);
-  const newQuestion = await  Question.findByIdAndUpdate(
-    id,
-    {subject,question},{runValidators:true,new:true}
-  );
-  console.log(newQuestion);
- res.redirect("/semester/newQuePage");
-}))
-
-//suggested question delete || delete request
-
-app.delete("/semester/newQues/:id",isQuestionOwner,wrapAsync(async(req,res)=>{
-  let {id} = req.params;
-  let deletedQues = await Question.findByIdAndDelete(id);
-  console.log(deletedQues);
-  res.redirect("/semester/newQuePage");
-}))
-
-
-
-//Sign-in route get/post
-app.get("/semester/signin",(req,res)=>{
-  res.render("users/signinForm");
-})
-
-//post
-app.post("/semester/signin",validateUser,wrapAsync(async (req,res)=>{
-  try{
-      let {username,email,password,role,semester} = req.body;
-      let newUser = new User({
-         username:username,
-         email:email,
-         role:role,
-         semester:semester
-     })
-  let user = await User.register(newUser,password);
-  req.login(user,(err)=>{
-    if(err){
-      return next(err);
-    }
-    req.flash("success","You Successfully sign-In, Welcome To Regal College!");
-    res.redirect("/");
-  })
-  
-  }catch(e){
-    req.flash("error",e.message);
-    res.redirect("/semester/signin");
-  }
-
-}));
-
-
-//login route get/post
-app.get("/semester/login",(req,res)=>{
-  res.render("users/loginForm")
-})
-
-
-app.post("/semester/login",saveReturnTo,passport.authenticate("local",{
-  failureRedirect:"/semester/login",
-  failureFlash:true,
-}),wrapAsync(async (req,res)=>{
-  req.flash("success","Welcome Back to regal College");
-  let redirectUrl = res.locals.redirectUrl || "/";
-  console.log(redirectUrl);
-  res.redirect(redirectUrl);
-}))
-
-
-app.get("/logout",(req,res)=>{
-  req.logOut((err)=>{
-    if(err){
-      return next(err);
-    }
-    req.flash("success","You Logged Out!");
-    res.redirect("/");
-  });
-});
-
-//Semester 
+///Semester 
 app.get("/semesters/:id", wrapAsync(async(req, res) => {
   const semId = parseInt(req.params.id);
   const Subjects = await Subject.find({semester: semId});
@@ -292,128 +106,25 @@ app.get("/semesters/:id", wrapAsync(async(req, res) => {
 
 
 
-//feedback form routes
-app.get("/feedback", isLoggedIn, (req,res) => {
-  res.render("feedback/feedback.ejs");
-});
+//Question Paper Routes
+app.use("/", questionPaperRoutes);
 
-app.post("/feedback",validateFeedback, isLoggedIn , wrapAsync(async (req,res) => {
-  const feedback = new Feedback(req.body);
-  feedback.user = req.user._id;
-  await feedback.save();
+//Suggested Question Routes
+app.use("/semester",suggestedQuestion);
 
-  const user = await User.findById(req.user._id);
-  user.feedbacks.push(feedback._id);
-  await user.save();
+//Task Routes
+app.use("/semester",taskRoutes);
 
-  req.flash("success", "Feedback submitted successfully!");
-  res.render("feedback/thankyou");
-}));
+//feedback Routes
+app.use("/feedback", feedbackRoutes);
 
+//subject Routes
+app.use("/semesters/:id/subjects",subjectRoutes);
 
-//feedback page route
-app.get("/feedbacks", async (req,res) => {
-  const allFeedbacks = await Feedback.find({})
-  .sort({ createdAt: -1})
-  .populate("user");
-
-  res.render("feedback/feedbackpage.ejs", {allFeedbacks});
-});
-
-//feedback delete route
-app.delete("/feedbacks/:id", async (req, res) => {
-  const { id } = req.params;
-  const feedback = await Feedback.findById(id);
-
-  if (!feedback) {
-    req.flash("error", "Feedback not found");
-    return res.redirect("/feedbacks");
-  }
-
-  await Feedback.findByIdAndDelete(id);
-  req.flash("success", "Feedback deleted successfully");
-  res.redirect("/feedbacks");
-});
+//User Routes
+app.use("/semester",userRoutes);
 
 
-
-//tasks page
-app.get("/semester/taskPage",isLoggedIn,wrapAsync(async (req,res)=>{
-  const task = await Task.find({}).populate("createdBy")
-  res.render("Tasks/taskPage.ejs",{task});
-}))
-
-
-//task Form
-app.get("/semester/task",isLoggedIn,isTeacher,(req,res)=>{
-  res.render("Tasks/taskForm.ejs");
-})
-
-
-//task post
-app.post("/semester/task",isLoggedIn,isTeacher,wrapAsync(async (req,res)=>{
-  let {title,description,semester} = req.body;
-  let createdBy = req.user._id;
-  let newTask = new Task({
-    title:title,
-    description:description,
-    semester:semester,
-    createdBy:createdBy
-  });
-  let currTask = await newTask.save();
-  console.log(currTask);
-  res.redirect("/semester/taskPage");
-}));
-
-//task edit
-app.get("/semester/task/:id/editForm",async (req,res)=>{
-  let {id} = req.params;
-  let task = await Task.findById(id).populate("createdBy");
-  res.render("Tasks/taskEditForm.ejs",{task});
-});
-
-//task update
-app.put("/semester/task/:id",async (req,res)=>{
-      let {id} = req.params;
-      let { title, description, semester} = req.body;
-
-      let newTask = await Task.findByIdAndUpdate(
-        id,
-        {title,description,semester},
-        {runValidators:true,new:true}
-      )
-      console.log(newTask);
-      res.redirect("/semester/taskPage");
-});
-
-//task delete
-app.delete("/semester/task/:id",async (req,res)=>{
-  let {id} = req.params;
-  let deletedTask = await Task.findByIdAndDelete(id);
-  console.log(deletedTask);
-  res.redirect("/semester/taskPage");
-});
-
-
-
-//subject routes
-
-app.get("/semesters/:id/subjects/new", isTeacher, (req, res) => {
-    const semester = parseInt(req.params.id);
-    res.render("subjects/new", { semester});
-});
-
-// POST: Create new subject for a semester
-app.post("/semesters/:id/subjects", isTeacher, async (req, res) => {
-    const semester = parseInt(req.params.id);
-    const { name } = req.body;
-
-    const subject = new Subject({ name, semester });
-    await subject.save();
-
-    req.flash("success", "Subject created successfully!");
-    res.redirect(`/semesters/${semester}`);
-});
 
 
 //page  not found
