+++ package.json
@@ -17,7 +17,7 @@
   "homepage": "https://github.com/Tamsil-sarnobat/question-bank-project#readme",
   "description": "",
   "dependencies": {
-    "cloudinary": "^2.7.0",
+    "cloudinary": "^1.41.3",
     "connect-flash": "^0.1.1",
     "cookie-parser": "^1.4.7",
     "dotenv": "^17.2.1",
