+++ views/feedback/feedback.ejs
@@ -20,76 +20,25 @@
       <div class="invalid-feedback">Please give your feedback.</div>
     </div>
 
-    <div class="mb-3">
-      <label class="form-label d-block">Rating:</label>
-      <div class="form-check form-check-inline">
-        <input
-          class="form-check-input"
-          type="radio"
-          name="rating"
-          id="rating1"
-          value="1"
-          required
-        />
-        <label class="form-check-label text-warning fs-4" for="rating1"
-          >★</label
-        >
-      </div>
-      <div class="form-check form-check-inline">
-        <input
-          class="form-check-input"
-          type="radio"
-          name="rating"
-          id="rating2"
-          value="2"
-          required
-        />
-        <label class="form-check-label text-warning fs-4" for="rating2"
-          >★★</label
-        >
-      </div>
-      <div class="form-check form-check-inline">
-        <input
-          class="form-check-input"
-          type="radio"
-          name="rating"
-          id="rating3"
-          value="3"
-          required
-        />
-        <label class="form-check-label text-warning fs-4" for="rating3"
-          >★★★</label
-        >
-      </div>
-      <div class="form-check form-check-inline">
-        <input
-          class="form-check-input"
-          type="radio"
-          name="rating"
-          id="rating4"
-          value="4"
-          required
-        />
-        <label class="form-check-label text-warning fs-4" for="rating4"
-          >★★★★</label
-        >
-      </div>
-      <div class="form-check form-check-inline">
-        <input
-          class="form-check-input"
-          type="radio"
-          name="rating"
-          id="rating5"
-          value="5"
-          required
-        />
-        <label class="form-check-label text-warning fs-4" for="rating5"
-          >★★★★★</label
-        >
-      </div>
 
-      <div class="invalid-feedback">Please give a rating.</div>
-    </div>
+<fieldset class="starability-slot mb-5">
+  <legend>Rating</legend>
+  <input type="radio" id="no-rate" class="input-no-rate" name="rating" value="0" checked aria-label="No rating." />
+  <input type="radio" id="first-rate1" name="rating" value="1" />
+  <label for="first-rate1" title="Terrible">1 star</label>
+  <input type="radio" id="first-rate2" name="rating" value="2" />
+  <label for="first-rate2" title="Not good">2 stars</label>
+  <input type="radio" id="first-rate3" name="rating" value="3" />
+  <label for="first-rate3" title="Average">3 stars</label>
+  <input type="radio" id="first-rate4" name="rating" value="4" />
+  <label for="first-rate4" title="Very good">4 stars</label>
+  <input type="radio" id="first-rate5" name="rating" value="5" />
+  <label for="first-rate5" title="Amazing">5 stars</label>
+  <div class="invalid-feedback">Please give a rating.</div>
+</fieldset>
+
+      
+
