+++ views/feedback/feedbackpage.ejs
@@ -34,7 +34,7 @@
 
               <% if (currentUser && fb.user && fb.user._id.toString() === currentUser._id.toString()) { %>
                 <form
-                  action="/feedbacks/<%= fb._id %>?_method=DELETE"
+                  action="/feedback/<%= fb._id %>?_method=DELETE"
                   method="POST"
                   class="d-inline"
                 >
