+++ views/includes/navbar.ejs
@@ -12,8 +12,8 @@
         <%if(teacher){%>   
         <a class="nav-link" href="/semester/task">Post Task</a>  
         <%}%>
-        <a class="nav-link" href="/feedback">feedback</a>           
-        <a class="nav-link" href="/logout"><b>Logout</b></a>     
+        <a class="nav-link" href="/feedback/new">feedback</a>           
+        <a class="nav-link" href="/semester/logout"><b>Logout</b></a>     
         <a class="nav-link" href="#" data-bs-toggle="modal" data-bs-target="#exampleModal"><img src="/Assets/userLogo.png" alt="user" class="userLogo"></a>  
        <% }%>
         <%if(!currUser){%>
